+++ pkgs/http_parser/lib/http_parser.dart
@@ -6,6 +6,7 @@ library http_parser;
 
 export 'src/authentication_challenge.dart';
 export 'src/case_insensitive_map.dart';
+export 'src/data_uri.dart';
 export 'src/http_date.dart';
 export 'src/media_type.dart';
 export 'src/web_socket.dart';